MNIST/requirements.txt
pillow
tqdm
torch
torchvision
fastapi
uvicorn
psutil
python-multipart
